добавление аватарки пользователем, сохранение её в базе данных, отображение всех созданных пользователями викторин при клике на кнопку старт => пользовательские

+++ src/quizBuild/quizBuildQuit.php
@@ -4,8 +4,11 @@ require_once __DIR__ . '/../helpers.php';
 
 if($_SERVER['REQUEST_METHOD'] == 'POST')
 {
+    if ($_SESSION['quiz_question_number'] !== 1) {
+        deleteQuiz();
+    }
+    
     deleteQuizName($_SESSION['quiz_name']);
-    deleteQuiz();
     unset($_SESSION['quiz_name']);
     unset($_SESSION['quiz_question_number']);
     redirect(path: '/profile.php');
